Separate penalties and score function
The scoring function would strip out common passwords when proceeding to
compute penalties. This would remove penalties from the original
password that we would like to keep.  This changes the function to retrieve
all the penalties from the original password, but it keeps the password score
computation by keeping the behavior of removing the common password from
the password in the score computation