add the default issuer in AWS 2FA QR code
When I add 2FA in AWS, the issuer would be "Amazon Web Services". Because the authenticator doesn't know this, it cannot find the icon. By this PR it can be fixed.
There are more missing default issuer, but because I'm not sure 100% about others, I haven't added them to the PR.